added buggy paging

+++ app/src/main/java/com/cythero/data/attraction/FakeAttractionRepository.kt
@@ -7,7 +7,7 @@ import com.cythero.domain.attraction.service.AttractionRepository
 @Suppress("RedundantNullableReturnType")
 object FakeAttractionRepository : AttractionRepository {
 
-	override suspend fun getMulti(page: Int, size: Int): AttractionHolder? = AttractionHolder.fake(size)
+	override suspend fun getMulti(page: Long, size: Long): AttractionHolder? = AttractionHolder.fake(size)
 
 	override suspend fun getOne(id: Long): Attraction? = Attraction.fake()
